I decided to put the DataObjectAttribute into custom-attributes for testing. attributes was changed to import it. Non primary was simply formatted
