+++ packages/markdoc-html-tags/src/lib/attributes.ts
@@ -1,6 +1,6 @@
-import type { Scalar, SchemaAttribute, ValidationError, ValidationType } from "@markdoc/markdoc";
+import type { Scalar, SchemaAttribute, ValidationType } from "@markdoc/markdoc";
 import {
-    MarkdocValidatorAttribute,
+    DataObjectAttribute,
     createAnArrayOfMarkdocErrorObjectsBasedOnEachConditionThatIsTrue,
     generateMarkdocErrorObject
 } from "src/utils";
@@ -182,47 +182,12 @@ export const contenteditable = generateBooleanAttributeSchemaThatIsNotRequired({
 });
 
 
-class DataObjectAttribute extends MarkdocValidatorAttribute {
-
-
-    override transform(value: Record<string, Scalar>,): Scalar {
-
-
-        const arrayTuplesWithKeysThatHaveDataAsThePrefixForEachWordAndIsCamelCased =
-            Object.entries(value).map(([key, value]) => [`data${key.at(0)}${key.substring(1, -1).toUpperCase()}`, value])
-
-        return Object.fromEntries(arrayTuplesWithKeysThatHaveDataAsThePrefixForEachWordAndIsCamelCased)
-
-
-    }
-
-
-    override returnMarkdocErrorObjectOrNothing(value: object,): ValidationError | void {
-
-        const regexToCheckIfAValueOnlyHasAlphanumericCharacters = /^[A-Za-z]+$/
-
-        const keysWithoutOnlyAlphanumericCharacters = Object.keys(value)
-            .filter(string => !regexToCheckIfAValueOnlyHasAlphanumericCharacters.test(string))
-
-        if (keysWithoutOnlyAlphanumericCharacters.length !== 0)
-            return generateMarkdocErrorObject(
-                "invalid-characters",
-                "error",
-                `These  are not good keys ${keysWithoutOnlyAlphanumericCharacters.join(",")}. 
-                They must be words with no spaces.
-                `
-            )
-
-
-    }
-
-}
 
 export const dataMarkdocAttributeSchema = getGenerateMarkdocAttributeSchema({
     type: DataObjectAttribute,
     description: "An attribute that allows an element's content to be editable",
     errorLevel: "critical",
-    required: false
+    required: false,
 })();
